+++ Model/Config.php
@@ -57,11 +57,14 @@ class Config
     public const XML_CONFIG_PATH_ENABLE_PRE_CHECKOUT_TRIGGER = 'payment/amwal_payments/enable_pre_checkout_trigger';
     public const XML_CONFIG_PATH_IS_PWA_MODE = 'payment/amwal_payments/pwa_mode';
     public const XML_CONFIG_PATH_ENABLE_BANK_INSTALLMENTS = 'payment/amwal_payments/enable_bank_installments';
+    public const XML_CONFIG_PATH_DISCOUNT_RULE = 'payment/amwal_payments/discount_rule';
+    public const XML_CONFIG_PATH_CARDS_BIN_CODES = 'payment/amwal_payments/cards_bin_codes';
+    public const XML_CONFIG_PATH_VIRTUAL_ITEMS_SUPPORT = 'payment/amwal_payments/virtual_items_support';
 
   /**
      * @var string
      */
-    const MODULE_VERSION = '1.0.35';
+    const MODULE_VERSION = '1.0.36';
 
     /** @var ScopeConfigInterface */
     private ScopeConfigInterface $scopeConfig;
@@ -531,5 +534,30 @@ class Config
             return trim($result->getOutput());
         }
         return '';
+    }
+
+    /**
+     * @return array
+     */
+    public function getCardsBinCodes(): array
+    {
+        $cardsBinCodes = $this->scopeConfig->getValue(self::XML_CONFIG_PATH_CARDS_BIN_CODES);
+        return $cardsBinCodes ? explode(',', $cardsBinCodes) : [];
+    }
+
+    /**
+     * @return string
+     */
+    public function getDiscountRule(): string
+    {
+        return (string) ($this->scopeConfig->getValue(self::XML_CONFIG_PATH_DISCOUNT_RULE) ?? '');
+    }
+
+    /**
+     * @return bool
+     */
+    public function isVirtualItemsSupport(): bool
+    {
+        return $this->scopeConfig->isSetFlag(self::XML_CONFIG_PATH_VIRTUAL_ITEMS_SUPPORT);
     }
 }